Move UI to another file
