+++ elephantfish.py
@@ -5,9 +5,6 @@ from __future__ import print_function
 import re, sys, time
 from itertools import count
 from collections import namedtuple
-from http.server import HTTPServer, SimpleHTTPRequestHandler
-import json
-from urllib.parse import parse_qs
 
 piece = { 'P': 44, 'N': 108, 'B': 23, 'R': 233, 'A': 23, 'C': 101, 'K': 2500}
 
@@ -377,16 +374,13 @@ class Searcher:
             lower, upper = -MATE_UPPER, MATE_UPPER
             while lower < upper - EVAL_ROUGHNESS:
                 gamma = (lower+upper+1)//2
-                print(f"in while: lower={lower}, upper={upper}, gamme={gamma}, depth={depth}")
                 score = self.bound(pos, gamma, depth)
-                print(f"in while: score={score}")
                 if score >= gamma:
                     lower = score
                 if score < gamma:
                     upper = score
             # We want to make sure the move to play hasn't been kicked out of the table,
             # So we make another call that must always fail high and thus produce a move.
-            print(f"out of while: lower={lower}, depth={depth}")
             self.bound(pos, lower, depth)
             # If the game hasn't finished we can retrieve our move from the
             # transposition table.
@@ -461,106 +455,5 @@ def main():
         print("Think depth: {} My move: {}".format(_depth, render(255-move[0] - 1) + render(255-move[1]-1)))
         hist.append(hist[-1].move(move))
 
-class ChessRequestHandler(SimpleHTTPRequestHandler):
-    def end_headers(self):
-        self.send_header('Access-Control-Allow-Credentials', 'true')
-        self.send_header('Access-Control-Allow-Origin', '*')
-        self.send_header('Access-Control-Allow-Methods', 'GET, POST, OPTIONS')
-        self.send_header("Access-Control-Allow-Headers", "X-Requested-With, Content-type")
-        SimpleHTTPRequestHandler.end_headers(self)
-
-    def do_OPTIONS(self):
-        print(f"get option req, path={self.path}")
-        self.send_response(200, "ok")
-        # self.send_header('Access-Control-Allow-Credentials', 'true')
-        # self.send_header('Access-Control-Allow-Origin', '*')
-        # self.send_header('Access-Control-Allow-Methods', 'GET, POST, OPTIONS')
-        # self.send_header("Access-Control-Allow-Headers", "X-Requested-With, Content-type")
-        self.end_headers()
-
-    def do_POST(self):
-        print(f"get post req, path={self.path}")
-        if self.path == '/move':
-            # 读取请求体
-            content_length = int(self.headers['Content-Length'])
-            post_data = self.rfile.read(content_length)
-            move_data = json.loads(post_data.decode('utf-8'))
-            
-            # 获取移动步骤
-            move = move_data.get('move')
-            if not move:
-                self.send_error(400, "Missing move parameter")
-                return
-            
-            try:
-                # 这里调用象棋引擎的代码处理移动并获取 AI 的响应
-                print(f"key log: To Process move: {move}")
-                ai_move = process_move(move)  # 这个函数需要实现
-                # print(f"key log: AI move: {ai_move}")
-                
-                # 发送响应
-                self.send_response(200)
-                self.send_header('Content-type', 'application/json')
-                self.end_headers()
-                self.wfile.write(json.dumps({'move': ai_move}).encode('utf-8'))
-            except Exception as e:
-                self.send_error(500, str(e))
-
-def run_server(port=8000):
-    server_address = ('', port)
-    httpd = HTTPServer(server_address, ChessRequestHandler)
-    print(f"Starting chess server on port {port}...")
-    httpd.serve_forever()
-
-hist = [Position(initial, 0)]
-searcher = Searcher()
-
-def process_move(input_move):
-    """
-    处理玩家的移动并返回 AI 的响应
-    :param input_move: 字符串，格式如 "h2e2"
-    :return: 字符串，AI 的移动，格式如 "h7h6"
-    """
-
-    print_pos(hist[-1])
-
-    if hist[-1].score <= -MATE_LOWER:
-        return ("You lost")
-
-    print(f"key log: input_move: {input_move}")
-    match = re.match('([a-i][0-9])'*2, input_move)
-    if not match:
-        return ("Please enter a move like h2e2") # taylor use error code
-    move = parse(match.group(1)), parse(match.group(2))
-    if move not in hist[-1].gen_moves():
-        return "ErrInvalidMove"
-
-    print(f"key log: input move is valid")
-    hist.append(hist[-1].move(move))
-
-    # After our move we rotate the board and print it again.
-    # This allows us to see the effect of our move.
-    print_pos(hist[-1].rotate())
-
-    if hist[-1].score <= -MATE_LOWER:
-        return ("You won")
-
-    # Fire up the engine to look for a move.
-    start = time.time()
-    for _depth, move, score in searcher.search(hist[-1], hist):
-        if time.time() - start > THINK_TIME:
-            break
-
-    if score == MATE_UPPER:
-        print("Checkmate!")
-
-    # The black player moves from a rotated position, so we have to
-    # 'back rotate' the move before printing it.
-    ai_move = render(255-move[0] - 1) + render(255-move[1]-1)
-    print("Think depth: {} My move: {}".format(_depth, ai_move))
-    hist.append(hist[-1].move(move))
-    print_pos(hist[-1])
-    return ai_move
-
 if __name__ == '__main__':
-    run_server()
+    main()
